requirements.txt
aniso8601==8.0.0
click==7.1.2
Flask==1.1.2
Flask-RESTful==0.3.8
Flask-SQLAlchemy==2.4.1
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
pkg-resources==0.0.0
psycopg2-binary==2.8.5
python-dotenv==0.13.0
pytz==2020.1
six==1.14.0
SQLAlchemy==1.3.16
Werkzeug==1.0.1
